Auton measurements & fixes

+++ FRC2017-test/src/DriveUntil.cpp
@@ -3,7 +3,7 @@
 #include "Globals.h"
 
 
-DriveUntil::DriveUntil(double x, double y, double z, std::vector<int> all) {
+DriveUntil::DriveUntil(double x, double y, double z, std::vector<double> all) {
 	this->x = x; this->y = y; this->z = z; this->untilAll = all;
 }
 
@@ -17,7 +17,7 @@ bool DriveUntil::periodic() {
 }
 
 bool DriveUntil::shouldStop() {
-	std::vector<int> encoderVals = theDrivetrain->getDistances();
+	std::vector<double> encoderVals = theDrivetrain->getDistances();
 	bool toReturn = true;
 	for (int i = 0; i != 4; ++i) {
 		if (encoderVals[i] - startingVals[i] < untilAll[i]) {
@@ -31,6 +31,6 @@ bool DriveUntil::shouldStop() {
 
 DriveUntil turnTo(double degrees) {
 	double turnAmount = degrees * DEG_TO_INCH_MULTIPLIER * INCH_TO_REV_MULTIPLIER;
-	return DriveUntil(0, 0, degrees > 0 ? 1 : -1, std::vector<int>(4, turnAmount));
+	return DriveUntil(0, 0, degrees > 0 ? 1 : -1, std::vector<double>(4, turnAmount));
 }
