+++ FRC2017-test/src/Subsystems/Drivetrain.h
@@ -25,7 +25,7 @@ public:
 	double GetEncoderDistance();
 	void ResetDistances();
 	void DriveWithStick(int facing);
-	std::vector<int> getDistances();
+	std::vector<double> getDistances();
 
 	constexpr static double TOP_SPEED = 100;
 };
